Fixed a bunch of minor bugs the program can work repeatedly even when option 2 is selected
I sorta-kinda-somewhat fixed a bug where if the user did not enter
a number length for the number generator it would crash. I put a
try-catch with an nextLine input thingy. It works pretty well for now
I want to improve it by making a method out of the try block but
its proving to be harder than anticapated.